app/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Mario Wanka <wanka.mario@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.1.1"
psycopg2-binary = "^2.9.10"
sqlalchemy = "^2.0.41"
cattrs = "^25.1.1"
python-dateutil = "^2.9.0.post0"
orjson = "^3.10.18"
flask-wtf = "^1.2.2"
dotenv = "^0.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
